requirements.txt
langchain
openai
pandas
sqlalchemy
python-dotenv
kaggle
serpapi
requests
tqdm
typeguard